Added the ability to calculate with decimal numbers

+++ app/src/main/java/com/onur/se2einzelphase/Calculator.java
@@ -9,8 +9,8 @@ public class Calculator {
     public Calculator() {
     }
 
-    public int divide(int dividend, int divisor){
-        int result=0;
+    public double divide(double dividend, double divisor){
+        double result=0;
         if(divisor==0){
             throw new IllegalArgumentException("Nicht möglich!");
         }
